move query_filter func
