+++ tests/blueprint/test_filters.py
@@ -2,7 +2,7 @@ import datetime as dt
 
 from cuenca_validations.types import QueryParams
 
-from agave.filters import generic_query
+from agave.collections.mongo.filters import generic_query
 
 
 def test_generic_query_before():
